requirements.txt
requests
gmail
wtforms
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
